agregando validaciones y login

+++ Backend/src/main/java/c19_65_t_java_js/NoCountry/SmartBank/model/Movimiento.java
@@ -4,6 +4,10 @@ package c19_65_t_java_js.NoCountry.SmartBank.model;
 
 import c19_65_t_java_js.NoCountry.SmartBank.DTO.MovimientoDTO;
 import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.Size;
 import lombok.*;
 
 import java.time.LocalDateTime;
@@ -13,22 +17,17 @@ import java.time.LocalDateTime;
 public class Movimiento {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+
     private Long id;
 
 //    @ManyToOne
 //    @JoinColumn(name = "idCuenta")
     private String usuario;
-
-    private Integer monto;
-
-    private Integer saldo;
-
+    private double monto;
+    private double saldo;
     private String ctaOrigen;
-
     private String ctaDestino;
-
     private String descripcionMovimiento;
-
     private LocalDateTime fechaMovimiento;
